# specs/002-m3-auth-thread-message/contracts/http-m3.openapi.yaml
openapi: 3.1.0
info:
  title: Loomi M3 Auth, Thread, and Message API
  version: 0.1.0
  description: Local-development API contract for M3 identity, threads, and user-authored messages.
servers:
  - url: http://127.0.0.1:8080
    description: Local Loomi API service
paths:
...
  /v1/threads:
    get:
      summary: List active threads for the fixed local identity
      operationId: listThreads
      parameters:
        - name: include_archived
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: When true, include archived durable threads. Default active list excludes archived threads.
      responses:
        '200':
          description: Thread list ordered by most recently updated first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadListResponse'
    post:
      summary: Create a thread for the fixed local identity
      operationId: createThread
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateThreadRequest'
      responses:
        '201':
          description: Created thread
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
...
components:
...
  responses:
    InvalidRequest:
      description: Invalid client request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            emptyMessage:
              value:
                error:
                  code: invalid_request
                  message: Message content is required.
                  request_id: req_000000000000000000000001
...
    MethodNotAllowed:
      description: HTTP method is not supported for this endpoint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unsupported:
              value:
                error:
                  code: method_not_allowed
                  message: Unsupported method.
                  request_id: req_000000000000000000000003
  schemas:
...
    ThreadListResponse:
      type: object
      required: [threads, request_id]
      properties:
        threads:
          type: array
          items:
            $ref: '#/components/schemas/Thread'
        request_id:
          type: string
    ThreadResponse:
      type: object
      required: [thread, request_id]
      properties:
        thread:
          $ref: '#/components/schemas/Thread'
        request_id:
          type: string
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum: [invalid_request, thread_not_found, method_not_allowed, internal_error]
            message:
              type: string
            request_id:
              type: string
    CreateThreadRequest:
      type: object
      required: [title, mode]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        mode:
          type: string
          enum: [chat, work]
      additionalProperties: false
...
    Thread:
      type: object
      required: [id, title, mode, lifecycle_status, created_at, updated_at]
      properties:
        id:
          type: string
        title:
          type: string
        mode:
          type: string
          enum: [chat, work]
        lifecycle_status:
          type: string
          enum: [active, archived]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        archived_at:
          type: [string, 'null']
          format: date-time
...